Update de.tsx
What are "Channels" (line 352, 347,  2445)? Channel in German is Kanal (like a TV channel), but I am not sure if this is the right translation in this case. Also what is meant by "Running..." in line 1674? Something like starting? Or does it mean something is running like a synchronisation? Thanks for al your amazing work!